Windows 79% unit test passing (#27)
* [DOCS]: ReadMe update

* [Windows]: Fixing [Windows build failer](https://github.com/kiliakis/BLonD-minimal-cpp/issues/25)

* [DOCS]: README.md bugfix

* [WINDOWS]: fixing 'long long' Warning!!! "Acceleration is not possible" errors (gone from 50% unit test sucsess to 79%)
The following tests FAILED:
          4 - testFullRing (Failed)
          6 - testInducedVoltage (Failed)
          7 - testInducedVoltageFreq (Failed)
          9 - testLHCFlatSpectrum (Failed)
         23 - testTC5_Wake_impedance (Failed)

+++ src/beams/Beams.cpp
@@ -9,7 +9,7 @@
 #include <blond/constants.h>
 #include <blond/math_functions.h>
 
-Beams::Beams(const uint _n_macroparticles, const long _intensity) {
+Beams::Beams(const uint _n_macroparticles, const long long _intensity) {
 
     this->n_macroparticles = _n_macroparticles;
     this->intensity = _intensity;
